More PR comments from CodeLens
