+++ test/mongoGetCommand.test.ts
@@ -5,8 +5,8 @@
 
 // The module 'assert' provides assertion methods from node
 import * as assert from 'assert';
-import { MongoCommands } from '../src/mongo/MongoCommands';
 import { Position } from 'vscode';
+import { getCommandFromText } from '../src/mongo/MongoScrapbook';
 
 function testParseExpectError(text: string, expected: { collection: string, name: string, args: object[] }) {
     let caughtError = false;
@@ -20,7 +20,7 @@ function testParseExpectError(text: string, expected: { collection: string, name
 }
 
 function testParse(text: string, expected: { collection: string, name: string, args: object[] }) {
-    let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+    let command = getCommandFromText(text, new Position(0, 0));
 
     assert.equal(command.collection, expected.collection, "Parsed collection name is not correct");
     assert.equal(command.name, expected.name, "Parsed command name is not correct");
@@ -32,13 +32,13 @@ function testParse(text: string, expected: { collection: string, name: string, a
 suite("scrapbook parsing Tests", () => {
     test("find", () => {
         let text = "db.find()";
-        let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+        let command = getCommandFromText(text, new Position(0, 0));
         assert.equal(command.text, text);
     });
 
     test("find with semicolon", () => {
         let text = "db.find();";
-        let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+        let command = getCommandFromText(text, new Position(0, 0));
         assert.equal(command.text, text);
     });
 
@@ -46,7 +46,7 @@ suite("scrapbook parsing Tests", () => {
         let line1 = "db.find()";
         let line2 = "db.insertOne({'a': 'b'})";
         let text = `${line1}\n${line2}`;
-        let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+        let command = getCommandFromText(text, new Position(0, 0));
         assert.equal(command.text, line1);
     });
 
@@ -54,7 +54,7 @@ suite("scrapbook parsing Tests", () => {
         let line1 = "db.find()";
         let line2 = "db.insertOne({'a':'b'})";
         let text = `${line1}\n${line2}`;
-        let command = MongoCommands.getCommandFromText(text, new Position(2, 0));
+        let command = getCommandFromText(text, new Position(2, 0));
         assert.equal(command.text, line2);
     });
 
@@ -62,7 +62,7 @@ suite("scrapbook parsing Tests", () => {
         let line1 = "db.find();";
         let line2 = "db.insertOne({'a':'b'})";
         let text = `${line1}\n${line2}`;
-        let command = MongoCommands.getCommandFromText(text, new Position(2, 0));
+        let command = getCommandFromText(text, new Position(2, 0));
         assert.equal(command.text, line2);
     });
 
@@ -70,7 +70,7 @@ suite("scrapbook parsing Tests", () => {
         let line1 = "db.find()";
         let line2 = "db.insertOne({'a': 'b'})";
         let text = `${line1}\r\n${line2}`;
-        let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+        let command = getCommandFromText(text, new Position(0, 0));
         assert.equal(command.text, line1);
     });
 
@@ -78,7 +78,7 @@ suite("scrapbook parsing Tests", () => {
         let line1 = "db.find()";
         let line2 = "db.insertOne({'a':'b'})";
         let text = `${line1}\r\n${line2}`;
-        let command = MongoCommands.getCommandFromText(text, new Position(2, 0));
+        let command = getCommandFromText(text, new Position(2, 0));
         assert.equal(command.text, line2);
     });
 
@@ -86,7 +86,7 @@ suite("scrapbook parsing Tests", () => {
         let line1 = "db.find()";
         let line2 = "db.insertOne({'a':'b'})";
         let text = `\r\n\r\n\r\n\r\n\r\n\r\n${line1}\r\n\r\n\r\n\r\n\r\n\r\n${line2}\r\n\r\n\r\n\r\n\r\n\r\n`;
-        let command = MongoCommands.getCommandFromText(text, new Position(5, 0));
+        let command = getCommandFromText(text, new Position(5, 0));
         assert.equal(command.text, line2);
     });
 
@@ -94,7 +94,7 @@ suite("scrapbook parsing Tests", () => {
         let line1 = "db.find()";
         let line2 = "db.insertOne({'a': 'b'})";
         let text = `${line1}\r\n\r\n\r\n${line2}`;
-        let command = MongoCommands.getCommandFromText(text, new Position(2, 0));
+        let command = getCommandFromText(text, new Position(2, 0));
         assert.equal(command.text, line1);
     });
 
@@ -196,7 +196,7 @@ suite("scrapbook parsing Tests", () => {
         let arg0 = `{"Age": 31}`;
         let arg1 = `{"Name": true}`;
         let text = `db.find(${arg0}\r\n,\r\n\r\n${arg1})`;
-        let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+        let command = getCommandFromText(text, new Position(0, 0));
         assert.deepEqual(JSON.parse(command.arguments![0]), JSON.parse(arg0));
         assert.deepEqual(JSON.parse(command.arguments![1]), JSON.parse(arg1));
     });
@@ -204,14 +204,14 @@ suite("scrapbook parsing Tests", () => {
         let arg0 = `[{"name": "a"}, {"name": "b"}, {"name": "c"}]`;
         let arg1 = `{"ordered": true}`;
         let text = `db.test1.insertMany(${arg0},\r\n\r\n\r\n${arg1})`;
-        let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+        let command = getCommandFromText(text, new Position(0, 0));
         assert.deepEqual(JSON.parse(command.arguments![0]), JSON.parse(arg0));
         assert.deepEqual(JSON.parse(command.arguments![1]), JSON.parse(arg1));
     });
     test("test function call that has a nested parameter", () => {
         let arg0 = `{"name": {"First" : "a", "Last":"b"} }`;
         let text = `db.test1.insertMany(${arg0})`;
-        let command = MongoCommands.getCommandFromText(text, new Position(0, 0));
+        let command = getCommandFromText(text, new Position(0, 0));
         assert.deepEqual(JSON.parse(command.arguments![0]), JSON.parse(arg0));
     });
 });
